feat: bundle CLI in Python wheel
