+++ .github/workflows/ci.yml
@@ -299,6 +299,7 @@ jobs:
           /tmp/kglite-networkx-smoke/bin/pip install "${wheel}[networkx]"
           cd /tmp
           /tmp/kglite-networkx-smoke/bin/python -c "import networkx as nx; import pandas; import kglite; graph = kglite.from_networkx(nx.path_graph(3)); assert graph.to_networkx().number_of_nodes() == 3"
+          /tmp/kglite-networkx-smoke/bin/kglite --help | grep -- 'code-tree'
           /tmp/kglite-networkx-smoke/bin/kglite-mcp-server --help | grep -- '--selftest'
 
       - name: Prefetch all-feature dependency metadata
