+++ kglite/__init__.pyi
@@ -706,6 +706,16 @@ def cypher_pass_names() -> list[str]:
     """
     ...
 
+def _run_cli(argv: list[str]) -> None:
+    """Run the bundled Rust ``kglite`` CLI in-process.
+
+    Internal entry point for the ``kglite`` console script (see
+    ``kglite/cli.py``). Not part of the public Python API; use the ``kglite``
+    command instead. ``argv`` excludes the program name and is parsed by clap
+    in the shared ``kglite-cli`` Rust library.
+    """
+    ...
+
 def _run_mcp_server(argv: list[str], embedder_factory: Optional[Callable[[str], EmbeddingModel]] = None) -> None:
     """Run the bundled `kglite-mcp-server` in-process; block until it exits.
